requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
pydantic[email]
PyJWT
websockets
sentence-transformers 
faiss-cpu 
rapidfuzz 
unidecode 
emoji